refactor: host partitionStoredRuleCapabilityErrors in shared
The bootstrap import of the helper from config tipped the capped
runtime-to-config layer exception (7/6). The pure helper moves to
shared/ (its right layer), settings-fleet-import re-exports it for
existing callers, and startup/restart-sync import the shared path.
Architecture, tsc and config suites green.

Ticket: Q-0146-3246
